Overwrite dynamic var with test var path
The `binding` context which properly sets the `*testing-path*` for the
fixtures and test vars is lost by the time `wrap-test-with-timer` is
called, which is why we're continuously reporting the `once` fixture
as the long running test.

To break out of this, we have to use `set!` to overwrite the dynamic
var. The downside to this is that we have to choose between reporting
on the `once` fixture, the `each` fixture, or the actual test var.

There is, presumably, a way to get accurate reporting for all of
these, but I'm not exactly sure how yet.